Split out ``test_ext_autodoc_typehints``
